reader: rename multi -> tailx
