+++ reader/tailx/tailx.go
@@ -1,4 +1,4 @@
-package multi
+package tailx
 
 import (
 	"errors"
